+++ tests/compas/data/test_dataschema.py
@@ -21,7 +21,7 @@ from compas.geometry import Torus
 from compas.geometry import Pointcloud
 
 from compas.datastructures import Graph
-from compas.datastructures import HalfEdge
+from compas.datastructures import Mesh
 
 if not compas.IPY:
     import jsonschema.exceptions
@@ -668,7 +668,7 @@ if not compas.IPY:
             Graph.validate_data(graph)
 
     @pytest.mark.parametrize(
-        "halfedge",
+        "mesh",
         [
             {
                 "attributes": {},
@@ -756,11 +756,11 @@ if not compas.IPY:
             },
         ],
     )
-    def test_schema_halfedge_valid(halfedge):
-        HalfEdge.validate_data(halfedge)
+    def test_schema_mesh_valid(mesh):
+        Mesh.validate_data(mesh)
 
     @pytest.mark.parametrize(
-        "halfedge",
+        "mesh",
         [
             {
                 "attributes": {},
@@ -812,12 +812,12 @@ if not compas.IPY:
             },
         ],
     )
-    def test_schema_halfedge_invalid(halfedge):
+    def test_schema_mesh_invalid(mesh):
         with pytest.raises(jsonschema.exceptions.ValidationError):
-            HalfEdge.validate_data(halfedge)
+            Mesh.validate_data(mesh)
 
     @pytest.mark.parametrize(
-        "halfedge",
+        "mesh",
         [
             {
                 "attributes": {},
@@ -845,6 +845,6 @@ if not compas.IPY:
             },
         ],
     )
-    def test_schema_halfedge_failing(halfedge):
+    def test_schema_mesh_failing(mesh):
         with pytest.raises(TypeError):
-            HalfEdge.validate_data(halfedge)
+            Mesh.validate_data(mesh)