+++ CHANGELOG.md
@@ -13,11 +13,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+* Merged `compas.datastructures.Halfedge` into `compas.datastructures.Mesh`.
 * Merged `compas.datastructures.Network` into `compas.datastructures.Graph`.
 
 ### Removed
 
 * Removed `compas.datastructures.Network`.
+* Removed `compas.datastructures.Halfedge`.
 
 ## [2.0.0-beta.2] 2024-01-12